Add 5 new tiled geo files for nationalmap.gov
Nationalmap.gov supports 256x256 tiles stored in a z/y/x format as
opposed to OSM's z/x/y. In addition there's no extension such as .png
on the tiles. These new geo files work with the source code mods
(previously committed) that add support for retrieving tiles stored
on the nationalmap.gov server. The maps are the same as those that
are available via WMS.